backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
dnspython==1.16.0
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==3.12.0
python-dotenv==1.1.1
repoze.lru==0.7
Routes==2.5.1
six==1.17.0
Werkzeug==3.1.3
